rbd: update remaining header fields for v2
There are three fields that are not yet updated for format 2 rbd
image headers:  the version of the header object; the encryption
type; and the compression type.  There is no interface defined
for the latter two, so just initialize them explicitly to 0 for
now.

We have several places we can get the version for the header object.
We know the snapshot context will be changing regularly, so use that
spot to get the version.  Have the caller supply the buffer into
which the version should be recorded.

Signed-off-by: Alex Elder <elder@inktank.com>