rbd: get snapshot name for a v2 image
Define rbd_dev_v2_snap_name() to fetch the name for a particular
snapshot in a format 2 rbd image.

Define rbd_dev_v2_snap_info() to to be a wrapper for getting the
name, size, and features for a particular snapshot.

Define rbd_dev_snap_info() as a wrapper to call the appropriate
function for getting the snapshot name, size, and features,
dependent on the rbd image format.

Signed-off-by: Alex Elder <elder@inktank.com>