rbd: encapsulate code that gets snapshot info
Create a function that encapsulates looking up the name, size and
features related to a given snapshot, which is indicated by its
index in an rbd device's snapshot context array of snapshot ids.

At the moment this (looking up the name anyway) is less efficient
than what's done currently, but we may be able to optimize this a
bit later on by cacheing the last lookup.

Signed-off-by: Alex Elder <elder@inktank.com>