rbd: define rbd_dev_v2_snapc_refresh()
Define a new function to update/refresh the snapshot context for a
format version 2 rbd image.  Update rbd_refresh_header() so it
selects which function to use based on the image format.

Unrelated--we use rbd_image_format_valid() here.  Delete the other
use of it, which was primarily put in place to ensure that function
was referenced at the time it was defined.

Signed-off-by: Alex Elder <elder@inktank.com>