rbd: two features masks are supplied, not one
The get_features rbd header object method returns two complete
features bitmasks--one indicating those supported by the remote end
and the second indicating which ones this end of the connection must
support for the connection to function.

We were only expecting one, but two were supplied.  This caused a
buffer overrun (which is a deeper problem we'll need to fix).

Signed-off-by: Alex Elder <elder@inktank.com>